+++ src/Shared/Recruit.Vacancies.Client/Domain/Repositories/IVacancyReviewRepository.cs
@@ -2,6 +2,7 @@
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using Esfa.Recruit.Vacancies.Client.Infrastructure.QueryStore.Projections.QA;
 
 namespace Esfa.Recruit.Vacancies.Client.Domain.Repositories
 {
@@ -12,7 +13,7 @@ namespace Esfa.Recruit.Vacancies.Client.Domain.Repositories
         Task<VacancyReview> GetAsync(Guid reviewId);
         Task UpdateAsync(VacancyReview review);
         Task<List<VacancyReview>> GetForVacancyAsync(long vacancyReference);
-        Task<List<VacancyReviewSearch>> SearchAsync(long vacancyReference);
+        Task<List<QaVacancySummary>> SearchAsync(long vacancyReference);
         Task<List<VacancyReview>> GetByStatusAsync(ReviewStatus status);
         Task<int> GetApprovedCountAsync(string submittedByUserId);
         Task<int> GetApprovedFirstTimeCountAsync(string submittedByUserId);
